fix: solve compilation errors caused by API changes in Smooks 2 (#21)

+++ src/test/java/org/smooks/cartridges/camel/processor/SmooksProcessor_JavaResult_Test.java
@@ -136,8 +136,8 @@ public class SmooksProcessor_JavaResult_Test extends CamelTestSupport {
         
         Coordinate coord = response.getMessage().getBody(Coordinate.class);
         
-        assertEquals(111, coord.getX());
-        assertEquals(222, coord.getY());
+        assertEquals((Integer) 111, coord.getX());
+        assertEquals((Integer) 222, coord.getY());
     }
 
 }
